refactor: moving TabsPagerAdapter to forder adapter
